pyproject.toml
[tool.poetry]
name = "sticker-bot"
version = "0.1.0"
description = ""
authors = ["Tetenkov Alexander <forthe@ya.ru>"]
readme = "README.md"
packages = [{include = "sticker_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
fastapi = "^0.99.0"
pillow = "^10.0.0"
uvicorn = "^0.22.0"
pydantic = "1.10.10"
python-dotenv = "^1.0.0"
aiogram = "3.0.0b7"
isort = "^5.12.0"
pytest = "^7.4.0"
httpx = "^0.24.1"
sqlitedict = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
mypy = "^1.4.1"
pre-commit = "^3.3.3"
types-pillow = "^9.5.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]
python_version = "3.11"
strict = true  # https://github.com/python/mypy/issues/11401
ignore_missing_imports = true
show_error_codes = true
allow_untyped_decorators = true
disallow_subclassing_any = false
implicit_reexport = true
exclude = ["venv", ".cache", "test_[a-z_]*.py", "conftest.py", "migrations", "fixtures.py"]
